# Conda prompt speedup and configuration