--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -10,6 +10,9 @@ POWERLEVEL9K_VCS_CLEAN_FOREGROUND='cyan'
 POWERLEVEL9K_VCS_CLEAN_BACKGROUND='white'
 ```
 
+Additionaly the vcs segment now has an `untracked` state which
+indicates that you have untracked files in your repository.
+
 The foreground color of actionformat is now configurable via:
 ```zsh
 POWERLEVEL9K_VCS_ACTIONFORMAT_FOREGROUND='green'