--- a/README.md
+++ b/README.md
@@ -140,18 +140,21 @@ The segments that are currently available are:
 
 ##### anaconda
 
-This segment shows your active anaconda environment. It relies on the
-`CONDA_ENV_PATH` environment variable to be set which happens when you properly
-`source activate` an environment.
+This segment shows your active anaconda environment. It relies on either the
+`CONDA_ENV_PATH` or the `CONDA_PREFIX` (depending on the `conda` version)
+environment variable to be set which happens when you properly `source
+activate` an environment.
 
+Special configuration variables:
 
 | Variable | Default Value | Description |
 |----------|---------------|-------------|
-|`POWERLEVEL9K_ANACONDA_BACKGROUND`|"green"|Background color of the segment.|
-|`POWERLEVEL9K_ANACONDA_FOREGROUND`|"black"|Foreground color of the segment.|
-|`POWERLEVEL9K_ANACONDA_PREFIX`|""|Prefix before the environment name.|
 |`POWERLEVEL9K_ANACONDA_LEFT_DELIMITER`|"("|The left delimiter just before the environment name.|
 |`POWERLEVEL9K_ANACONDA_RIGHT_DELIMITER`|")"|The right delimiter just after the environment name.|
+
+Additionally the following segment specific parameters can be used to customize
+it: `POWERLEVEL9K_PYTHON_ICON`, `POWERLEVEL9K_ANACONDA_BACKGROUND`, and
+`POWERLEVEL9K_ANACONDA_FOREGROUND`.
 
 ##### aws