--- a/powerlevel9k.zsh-theme
+++ b/powerlevel9k.zsh-theme
@@ -291,13 +291,8 @@ CURRENT_BG='NONE'
 
 # Anaconda Environment
 prompt_anaconda() {
-  if $(hash ack 2>/dev/null); then
-    local active_conda_env=$(where conda | ack -o '(?<=envs/)[\w-]+(?=/bin)')
-  else
-    local active_conda_env=$(where conda | grep -o -P '(?<=envs/)[\w-]+(?=/bin)')
-  fi
-  if [[ -n $active_conda_env ]]; then
-    "$1_prompt_segment" "$0" "$2" "green" "black" "($active_conda_env)" ""
+  if ! [ -z ${CONDA_ENV_PATH+x} ]; then
+    "$1_prompt_segment" "$0" "$2" "green" "black" "($(basename $CONDA_ENV_PATH))" ""
   fi
 }