This PR makes the (Ana)conda prompt much faster and fully configurable. It now parses the `CONDA_ENV_PATH` variable which ends up much faster:

```bash
# Old
$ time (where conda | ack -o '(?<=envs/)[\w-]+(?=/bin)')
0.04s user 0.01s system 94% cpu 0.051 total

# New
$ time (if ! [ -z ${CONDA_ENV_PATH+x} ]; then basename $CONDA_ENV_PATH; fi)
0.00s user 0.00s system 85% cpu 0.002 total
```

This might not seem like a lot but a lot of active segments already make for a sluggish feeling prompt so every bit helps.

It now also has the following config values:

| Variable | Default Value | Description |
|----------|---------------|-------------|
|`POWERLEVEL9K_ANACONDA_BG_COLOR`|"green"|Background color of the segment.|
|`POWERLEVEL9K_ANACONDA_FG_COLOR`|"black"|Foreground color of the segment.|
|`POWERLEVEL9K_ANACONDA_PREFIX`|""|Prefix before the environment name.|
|`POWERLEVEL9K_ANACONDA_LEFT_DELIMITER`|"("|The left delimiter just before the environment name.|
|`POWERLEVEL9K_ANACONDA_RIGHT_DELIMITER`|")"|The right delimiter just after the environment name.|

Thus one can go from this <img width="154" alt="screen shot 2016-07-15 at 17 35 41" src="https://cloud.githubusercontent.com/assets/800487/16879819/9fd8e1f2-4ab2-11e6-8571-dc83d0ecc0f5.png"> to for example this <img width="160" alt="screen shot 2016-07-15 at 17 38 40" src="https://cloud.githubusercontent.com/assets/800487/16879928/087f0bd2-4ab3-11e6-9fb7-19ddf0b62ff7.png"> and it is always nice to have options.