@dritter 2016-07-16
Cool thing! Thanks.

Just a few thoughts:

- Our development branch is `next`. Could you rebase your PR onto that branch (or just submit a new PR from that branch).

- You don't need explicit variables for foreground/background color, as the base `left/right_prompt_segment` got you already covered there. Every segment that uses that function has the ability to get it's colors overridden by defining a variable named `POWERLEVEL9K_<segment_name>_<state>_BACKGROUND`. "state" only if the segment has special states, of course.

- Setting default values (as you did for the delimiter) could be done by using our `set_default` function. Saves a bit of typing..

- You set the icon in `POWERLEVEL9K_ANACONDA_PREFIX`, right? I would use that as "visual identifier", which is the last parameter of `$1_prompt_segment`. The parameter should be a key of the icons array (defined in `functions/icons.zsh`; `PYTHON_ICON`, IMHO). The nice thing about that is, that the icon switches sides, if you display the segment on the left or the right. And users are still able to not display that icon by setting `POWERLEVEL9K_PYTHON_ICON=''`.