@bhilburn 2016-07-17
@krischer - Wow! Great work! This is going to be a really great contribution.

@dritter's feedback is good. Let us know if you need any help incorporating his suggestions. Once the PR is ready, I'm excited to get this merged.