@krischer 2016-07-19
Ah nice 😄  I totally missed the "Stylizing Your Prompt" section in the wiki. I'll act on all your suggestions - it might be until the weekend until I find the time though.

The example snake icon is an emoticon which not all terminals will be able to display. I'll look into another default choice.