@krischer 2016-08-30
Oops...I totally forgot about that - I've been happily using it after all ;-) You'll have it in a couple of minutes!